// apps/backend/src/model/kanel/public/ManifestDocument.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { ManifestId } from './Manifest';
import type { DocumentId } from './Document';

/** Represents the table public.Manifest_Document */
export default interface ManifestDocument {
  manifest_id: ManifestId;

  document_id: DocumentId;
}

/** Represents the initializer for the table public.Manifest_Document */
export interface ManifestDocumentInitializer {
  manifest_id: ManifestId;

  document_id: DocumentId;
}

/** Represents the mutator for the table public.Manifest_Document */
export interface ManifestDocumentMutator {
  manifest_id?: ManifestId;

  document_id?: DocumentId;
}