// apps/backend/src/model/kanel/public/ManifestRebuildQueue.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { ManifestType } from '../../../__generated__/resolvers-types';
import type { ManifestRebuildQueueStatus } from '../../../modules/shareable-resource/manifest/manifest.consts';

/** Identifier type for public.ManifestRebuildQueue */
export type ManifestRebuildQueueId = string & { __brand: 'public.ManifestRebuildQueue' };

/** Represents the table public.ManifestRebuildQueue */
export default interface ManifestRebuildQueue {
  id: ManifestRebuildQueueId;

  product: string;

  version: string;

  type: ManifestType;

  status: ManifestRebuildQueueStatus;

  created_at: Date;
}

/** Represents the initializer for the table public.ManifestRebuildQueue */
export interface ManifestRebuildQueueInitializer {
  /** Default value: gen_random_uuid() */
  id?: ManifestRebuildQueueId;

  product: string;

  version: string;

  type: ManifestType;

  status?: ManifestRebuildQueueStatus;

  /** Default value: CURRENT_TIMESTAMP */
  created_at?: Date;
}

/** Represents the mutator for the table public.ManifestRebuildQueue */
export interface ManifestRebuildQueueMutator {
  id?: ManifestRebuildQueueId;

  product?: string;

  version?: string;

  type?: ManifestType;

  status?: ManifestRebuildQueueStatus;

  created_at?: Date;
}